Separate `UtilTests.py` into individual files
